Add files via upload
I have a soft spot for figuring out algorithms so I decided to work with them and find a few that works the best. I first started making algorithms in Python in my first year in 2020, but moved to Java. This code is still in development but will be updated once I get the time as I am taking an algorithm class.