Add files via upload
First time making a binary tree and learning how it works. I have made more and tried to create a "Circular Step Binary Tree" but I haven't figured out how to create it yet. Idea being to be able to traverse through the tree using a range of numbers. Ex: First step of the tree is ranges 0-1,000 and second step is 1,001-2000, etc. That way if we want to traverse through a tree and find a number, we can search in a range and not have to traverse through all the numbers (depending on which type of search you use).